Move genric composes to new file
