+++ app/src/main/java/com/example/movieapp3/common/CoroutineScopeDispatchers.kt
@@ -1,12 +1,8 @@
 package com.example.movieapp3.common
 
-import com.example.movieapp3.listing.data.remote.MoviesRemoteDataSource
-import com.example.movieapp3.listing.data.remote.MoviesRemoteDataSourceImpl
-import dagger.Binds
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
-import dagger.hilt.android.components.ViewModelComponent
 import dagger.hilt.components.SingletonComponent
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.Dispatchers
