+++ app/src/main/java/com/example/movieapp3/common/presentation/GeneralComposables.kt
@@ -1,0 +1,53 @@
+package com.example.movieapp3.common.presentation
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.wrapContentSize
+import androidx.compose.material3.Button
+import androidx.compose.material3.CircularProgressIndicator
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.unit.dp
+import com.example.movieapp3.R
+
+@Composable
+fun RetryItem(retryAction: ()-> Unit) {
+    Column(
+        modifier = Modifier
+            .fillMaxWidth()
+            .padding(8.dp),
+        verticalArrangement = Arrangement.Center,
+        horizontalAlignment = Alignment.CenterHorizontally
+
+    ) {
+        Text(text = stringResource(R.string.msg_something_wrong))
+        Button(
+            modifier = Modifier
+                .wrapContentSize()
+                .padding(start = 20.dp, top = 8.dp, end = 20.dp, bottom = 8.dp),
+            onClick = retryAction
+        ) {
+            Text(text = stringResource(R.string.title_retry), color = Color.White)
+        }
+    }
+}
+
+@Composable
+fun CircularLoading() {
+    Column(
+        modifier = Modifier
+            .fillMaxWidth()
+            .padding(8.dp),
+        verticalArrangement = Arrangement.Center,
+        horizontalAlignment = Alignment.CenterHorizontally
+    ) {
+        CircularProgressIndicator()
+    }
+}
